refactor (WinningStatusCalculator) : 결합도 완화를 위해 추상타입을 구현하도록 변경
